refactor(hooks): Improve type safety and error handling in asset-related hooks
- Remove unnecessary `as const` type assertions in ABIs
- Update token logo import for XDC network
- Add type casting and null checks in useAssetDetails hook
- Improve type handling for asset price and reserve data
- Enhance error resilience in asset-related data fetching
- Simplify type annotations and reduce potential type errors
- Update asset details and price calculation logic

+++ src/hooks/useAssetPrice.ts
@@ -38,7 +38,9 @@ export const useAssetPrice = (assetAddress: string) => {
   });
 
   // Aave price oracle returns prices in USD with 8 decimals
-  const priceInUsd = assetPrice ? parseFloat(formatUnits(assetPrice, 8)) : 0;
+  const priceInUsd = assetPrice
+    ? parseFloat(formatUnits(assetPrice as bigint, 8))
+    : 0;
 
   return {
     price: priceInUsd,
